requirements.txt
django==2,<3.0
psycopg2>=2.7,<3.0
django-rest-framework
ipython
ipdb
